+++ src/payment/Payment.ts
@@ -1,5 +1,7 @@
+import Decimal from "decimal.js"
 import { Currency } from "../common"
-import { PaymentItem, Price } from "./PaymentItem"
+import { PaymentItem, paymentItemQuery } from "./PaymentItem"
+import { Price, priceMutation, priceQuery } from "./Price"
 import { TaxRate } from "./TaxRate"
 
 function getTaxRates(payment: Payment): Set<TaxRate> {
@@ -10,20 +12,34 @@ function getTaxRates(payment: Payment): Set<TaxRate> {
    return taxRates
 }
 
-function getTotalPrice(payment: Payment): Price {
-   const totalPrice: Price = {
-      whole: 0,
-      decimal: 0,
-   }
+function getPriceWithoutTax(payment: Payment): Price {
+   return getPrice(payment, paymentItemQuery.getPriceWithoutTax)
+}
+
+function getPriceWithTax(payment: Payment): Price {
+   return getPrice(payment, paymentItemQuery.getPriceWithTax)
+}
+
+function getPrice(
+   payment: Payment,
+   itemFn: (
+      item: PaymentItem,
+      opts?: {
+         roundCents?: boolean
+      },
+   ) => Price,
+): Price {
+   let priceCents = new Decimal(0)
    for (const [item, quantity] of payment.items) {
-      totalPrice.whole += item.price.whole * quantity
-      totalPrice.decimal += item.price.decimal * quantity
+      const itemPrice = itemFn(item, {
+         roundCents: false,
+      })
+      const itemPriceCents = new Decimal(
+         priceQuery.getAsCents(itemPrice, { roundCents: false }),
+      )
+      priceCents = priceCents.plus(itemPriceCents.times(quantity))
    }
-
-   totalPrice.whole += Math.floor(totalPrice.decimal / 100)
-   totalPrice.decimal %= 100
-
-   return totalPrice
+   return priceMutation.fromCents(priceCents.toNumber())
 }
 
 function getItemsWithDefaultTaxRate(
@@ -61,8 +77,9 @@ type Payment = {
 export type { Payment }
 export const paymentQuery = {
    getTaxRates,
-   getTotalPrice,
    getItemsWithDefaultTaxRate,
+   getPriceWithoutTax,
+   getPriceWithTax,
 }
 
 export const paymentMutation = {