+++ src/payment/TaxRate.ts
@@ -1,3 +1,5 @@
+// A number between 0 and 100
+
 type TaxRate = {
    name: string
    rate: number