+++ src/tests/payment/Payer.test.ts
@@ -1,11 +1,26 @@
 import { describe, expect, it } from "vitest"
-import { isVatValid } from "../../invoicing/checkVat"
+import { Payer, payerQuery } from "../../payment/Payer"
+
+function payerWithVat(vat: string): Payer {
+   return {
+      billingInfo: {
+         city: "Bratislava",
+         country: "Slovakia",
+         zip: "82108",
+         street: "Karlova Ves 123",
+      },
+      name: "John Doe",
+      companyInfo: {
+         businessId: "2020382056",
+         vatId: vat,
+      },
+   }
+}
 
 describe("Check VAT", () => {
    it("Should pass with valid VAT", async () => {
       const VAT = "SK2020382056"
-      const isValid = await isVatValid(VAT)
-      console.log(isValid)
+      const isValid = await payerQuery.hasValidVat(payerWithVat(VAT))
       expect(isValid.isErr()).toBe(false)
       if (isValid.isOk()) {
          expect(isValid.value).toBe(true)
@@ -14,7 +29,7 @@ describe("Check VAT", () => {
 
    it("Should pass with foreign VAT", async () => {
       const VAT = "FR63382357721" // psg vat
-      const isValid = await isVatValid(VAT)
+      const isValid = await payerQuery.hasValidVat(payerWithVat(VAT))
       console.log(isValid)
       expect(isValid.isErr()).toBe(false)
       if (isValid.isOk()) {
@@ -24,8 +39,8 @@ describe("Check VAT", () => {
 
    it("Should fail without prefix", async () => {
       const VAT = "2020382056"
-      const isValid = await isVatValid(VAT)
-      expect(isValid.isErr()).toBe(true)
+      const isValid = await payerQuery.hasValidVat(payerWithVat(VAT))
+      expect(isValid.isErr()).toBe(false)
       if (isValid.isOk()) {
          expect(isValid.value).toBe(false)
       }
@@ -33,8 +48,7 @@ describe("Check VAT", () => {
 
    it("Should fail with invalid VAT", async () => {
       const VAT = "SK2020382057"
-      const isValid = await isVatValid(VAT)
-      console.log(isValid)
+      const isValid = await payerQuery.hasValidVat(payerWithVat(VAT))
       expect(isValid.isErr()).toBe(false)
       if (isValid.isOk()) {
          expect(isValid.value).toBe(false)