+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 # @flagg2/payment
 
+## 0.0.9
+
+### Patch Changes
+
+-  add coercion to quantity
+
+## 0.0.8
+
+### Patch Changes
+
+-  add coercion to payment item
+
+## 0.0.7
+
+### Patch Changes
+
+-  add coercion to variableSymbol
+
+## 0.0.6
+
+### Patch Changes
+
+-  coerce dates
+
+## 0.0.5
+
+### Patch Changes
+
+-  accepts arrays in parseUnknown
+
 ## 0.0.4
 
 ### Patch Changes